Rename some classes to avoid confusion
The EventPickingActivityModule was renamed EventPickingModule, as the module contains more than just an activity.
The EventListModel was renamed to EventPickingModel, as it is a model for the whole module, and might do more than just fetching the list in the future.
The DaggerViewModelFactory was renamed ViewModelFactory to avoid confusion with Dagger generated classes that are prefixed by the Dagger name.