Set TimeLineAdapter class to properly integrate Concert objects
The Concert objects were transformed into a list of TimeLineModel which
contain the infos about the Concert.
Getters for each Concert field added in TimeLineModel.

These objects are put into TimeLineViewHolder objects.
TimeLineViewHolder attributes were linked to TextViews in layout files.

The TimeLineAdapter is the View equivalent (in the layout) to the List
containing the Concert objects. So it parses the TimeLineModel (which
are the transformed Concerts) into TimeLineViewHolder which will be
displayed.
Handling of this parsing was implemented in TimeLineAdapter.

Finally the ids of the views in the layout were renamed accordingly for
the binding.