Make Session and InMemorySession injectable
This commit takes the currently static Session methods and make them non-static. It also
makes the Session class injectable, as well as its inner component InMemorySession. The
DummyInMemorySession class is moved in a common test package and used directly in test/
and androidTest/, as it's not needed in the main code. All the tests and classes have
been updated accordingly.